#121: massive refactor move token types to lex and ParseTree to parse
