+++ fishi/templates/frontend.go.tmpl
@@ -16,7 +16,7 @@ import (
 
 	"github.com/dekarrin/ictiobus"
 	"github.com/dekarrin/ictiobus/trans"
-	"github.com/dekarrin/ictiobus/types"
+	"github.com/dekarrin/ictiobus/lex"
 {{if .IRPackage}}
     "{{ .IRPackage }}"
 {{end}}
@@ -75,7 +75,7 @@ func Frontend[IRType any](hooks trans.HookMap, opts *FrontendOptions) ictiobus.F
     // Add traces if requested
 
     if opts.LexerTrace {
-        fe.Lexer.RegisterTokenListener(func(t types.Token) {
+        fe.Lexer.RegisterTokenListener(func(t lex.Token) {
 			fmt.Fprintf(os.Stderr, "Token: %s\n", t)
 		})
     }
