+++ fishi/templates/tokens.go.tmpl
@@ -13,7 +13,6 @@ created by invoking ictiobus with the following command:
 
 import (
     "github.com/dekarrin/ictiobus/lex"
-    "github.com/dekarrin/ictiobus/types"
 )
 
 var (
@@ -25,7 +24,7 @@ var (
 {{end -}}
 )
 
-var all = map[string]types.TokenClass{
+var all = map[string]lex.TokenClass{
 {{range .Classes -}}
     "{{.ID}}": {{.Name}},
 {{end -}}
@@ -33,6 +32,6 @@ var all = map[string]types.TokenClass{
 
 // ByID returns the TokenClass in {{.Lang}} that has the given ID. If no token
 // class with that ID exists, nil is returned.
-func ByID(id string) types.TokenClass{
+func ByID(id string) lex.TokenClass{
     return all[id]
 }
