+++ lex/immediate_test.go
@@ -4,15 +4,13 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/dekarrin/ictiobus/types"
-
 	"github.com/stretchr/testify/assert"
 )
 
 func Test_ImmediateLex_singleStateLex(t *testing.T) {
 	testCases := []struct {
 		name       string
-		classes    []types.TokenClass
+		classes    []TokenClass
 		patterns   []string
 		lexActions []Action
 		input      string
@@ -52,7 +50,7 @@ func Test_ImmediateLex_singleStateLex(t *testing.T) {
 				{line: "someVar = (8 + 1)* 2", lineNum: 1, linePos: 17, class: testClassRParen, lexed: ")"},
 				{line: "someVar = (8 + 1)* 2", lineNum: 1, linePos: 18, class: testClassMult, lexed: "*"},
 				{line: "someVar = (8 + 1)* 2", lineNum: 1, linePos: 20, class: testClassInt, lexed: "2"},
-				{line: "someVar = (8 + 1)* 2", lineNum: 1, linePos: 21, class: types.TokenEndOfText},
+				{line: "someVar = (8 + 1)* 2", lineNum: 1, linePos: 21, class: TokenEndOfText},
 			},
 		},
 		{
@@ -89,7 +87,7 @@ func Test_ImmediateLex_singleStateLex(t *testing.T) {
 				{line: "someVar=(8+1)*2", lineNum: 1, linePos: 13, class: testClassRParen, lexed: ")"},
 				{line: "someVar=(8+1)*2", lineNum: 1, linePos: 14, class: testClassMult, lexed: "*"},
 				{line: "someVar=(8+1)*2", lineNum: 1, linePos: 15, class: testClassInt, lexed: "2"},
-				{line: "someVar=(8+1)*2", lineNum: 1, linePos: 16, class: types.TokenEndOfText},
+				{line: "someVar=(8+1)*2", lineNum: 1, linePos: 16, class: TokenEndOfText},
 			},
 		},
 		{
@@ -126,7 +124,7 @@ func Test_ImmediateLex_singleStateLex(t *testing.T) {
 				{line: "(8 + 1)* 2", lineNum: 2, linePos: 7, class: testClassRParen, lexed: ")"},
 				{line: "(8 + 1)* 2", lineNum: 2, linePos: 8, class: testClassMult, lexed: "*"},
 				{line: "(8 + 1)* 2", lineNum: 2, linePos: 10, class: testClassInt, lexed: "2"},
-				{line: "(8 + 1)* 2", lineNum: 2, linePos: 11, class: types.TokenEndOfText},
+				{line: "(8 + 1)* 2", lineNum: 2, linePos: 11, class: TokenEndOfText},
 			},
 		},
 	}
@@ -172,7 +170,7 @@ func Test_ImmediateLex_singleStateLex(t *testing.T) {
 				expectToken := tc.expect[tokNum]
 				actualToken := stream.Next()
 
-				if actualToken.Class().ID() == types.TokenError.ID() {
+				if actualToken.Class().ID() == TokenError.ID() {
 					assert.Fail("received error token", "error: %s", actualToken.Lexeme())
 				}
