+++ lex/stream.go
@@ -1,4 +1,4 @@
-package types
+package lex
 
 // TokenStream is a stream of tokens read from source text. The stream may be
 // lazily-loaded or immediately available.
