+++ parse/lalr_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/dekarrin/ictiobus/grammar"
-	"github.com/dekarrin/ictiobus/types"
+	"github.com/dekarrin/ictiobus/lex"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -92,7 +92,7 @@ func Test_LALR1Parse(t *testing.T) {
 				T -> T * F | F ;
 				F -> ( E ) | id ;
 			`,
-			input: []string{"(", "id", "+", "id", ")", "*", "id", types.TokenEndOfText.ID()},
+			input: []string{"(", "id", "+", "id", ")", "*", "id", lex.TokenEndOfText.ID()},
 			expect: `( E )
   \---: ( T )
           |---: ( T )
