+++ parse/parse.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/dekarrin/ictiobus/grammar"
 	"github.com/dekarrin/ictiobus/internal/box"
+	"github.com/dekarrin/ictiobus/lex"
 	"github.com/dekarrin/ictiobus/types"
 )
 
@@ -22,7 +23,7 @@ type Parser interface {
 
 	// Parse parses input text and returns the parse tree built from it, or a
 	// SyntaxError with the description of the problem.
-	Parse(stream types.TokenStream) (types.ParseTree, error)
+	Parse(stream lex.TokenStream) (ParseTree, error)
 
 	// Type returns a string indicating what kind of parser was generated. This
 	// will be "LL(1)", "SLR(1)", "CLR(1)", or "LALR(1)"
