+++ parse/tree_test.go
@@ -1,4 +1,4 @@
-package types
+package parse
 
 import (
 	"testing"
